src/backend/requirements.txt
# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.1

# Validation
pydantic>=2.10.0
email-validator>=2.2.0

# CORS
# fastapi-cors is deprecated, use fastapi.middleware.cors instead

# Utilities
python-dateutil>=2.9.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-env==1.1.3
pytest-mock==3.12.0
